Use flag

+++ main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"bytes"
+	"flag"
 	"fmt"
 	"go/ast"
 	"go/format"
@@ -13,6 +14,7 @@ import (
 	"os"
 	"path/filepath"
 	"reflect"
+	"runtime"
 	"sort"
 	"strings"
 
@@ -46,66 +48,43 @@ func getCustomTypeIface() *types.Interface {
 
 var customTypeIface = getCustomTypeIface()
 
-var config struct {
+var config = struct {
 	// AllowUnsafe enables the use of unsafe conversions.
-	AllowUnsafe bool
+	AllowUnsafe *bool
 
 	// CopyBytes disables the use of zero-copy byte slices.
-	CopyBytes bool
+	CopyBytes *bool
 
 	// AllowZeroCopyString enables the use of zero-copy string conversions.
 	// AllowUnsafe must be enabled for this to work.
-	AllowZeroCopyString bool
+	AllowZeroCopyString *bool
 
-	// Path
+	Version *bool
+
+	// Path (Argument)
 	Path string
+}{
+	AllowUnsafe:         flag.Bool("allow-unsafe", false, "Allow unsafe conversions"),
+	CopyBytes:           flag.Bool("copy-bytes", false, "Disable zero-copy bytes conversion"),
+	AllowZeroCopyString: flag.Bool("allow-zero-copy-string", false, "Enable zero-copy string conversion"),
+	Version:             flag.Bool("version", false, "Print version and exit"),
 }
 
-func Usage() {
-	fmt.Fprintln(os.Stderr, "Usage: gobe [options] <path>")
-	fmt.Fprintln(os.Stderr, "Options:")
-	fmt.Fprintln(os.Stderr, "  --allow-unsafe    Allow unsafe conversions")
-	fmt.Fprintln(os.Stderr, "  --copy-bytes    Disable zero-copy bytes conversion")
-	fmt.Fprintln(os.Stderr, "  --allow-zero-copy-string    Enable zero-copy string conversion")
-}
-
-func ParseParams(args []string) {
-	if len(args) == 0 {
-		Usage()
-		os.Exit(1)
-		return
-	}
-
-	for i := 0; i < len(args)-1; i++ {
-		arg := args[i]
-		var name string
-		if strings.HasPrefix(arg, "--") {
-			name = arg[2:]
-		} else {
-			break
-		}
-
-		switch name {
-		case "allow-unsafe":
-			config.AllowUnsafe = true
-		case "copy-bytes":
-			config.CopyBytes = true
-		case "allow-zero-copy-string":
-			config.AllowZeroCopyString = true
-		default:
-			log.Fatalf("unknown argument: %s", arg)
-		}
-	}
-
-	path := args[len(args)-1]
-	if path == "" {
-		log.Fatalf("path is empty")
-	}
-	config.Path = path
-}
+const VERSION = "0.0.5"
+const VERSION_STRING = "gobe version v" + VERSION + " " + runtime.GOOS + "/" + runtime.GOARCH
 
 func main() {
-	ParseParams(os.Args[1:])
+	flag.Parse()
+	if *config.Version {
+		fmt.Println(VERSION_STRING)
+		os.Exit(0)
+	}
+
+	if flag.NArg() != 1 {
+		flag.Usage()
+		os.Exit(1)
+	}
+	config.Path = flag.Arg(0)
 
 	os.Remove(filepath.Join(config.Path, "gobe_generated.go"))
 
@@ -222,7 +201,7 @@ func main() {
 	}
 
 	// Write to file
-	f, err := os.Create(filepath.Join(os.Args[1], "gobe_generated.go"))
+	f, err := os.Create(filepath.Join(config.Path, "gobe_generated.go"))
 	if err != nil {
 		panic(err)
 	}
